index is now search, new index, added searching
Rationality is index is the primary webpage to search whatever the user wants to buy.

Once the user searches, the form passes to searching where it essentially redirects to the dynamic URL for search.html.

Search.html now has a dynamic URL path of the actual query and displays query in webpage.